sprint 11 fixed

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,8 +2,10 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.util.List;
 
@@ -12,6 +14,7 @@ import java.util.List;
 @Service
 public class FilmService {
     private final FilmStorage storage;
+    private final UserStorage userStorage;
 
     public List<Film> findAll() {
         return storage.findAll();
@@ -26,10 +29,16 @@ public class FilmService {
     }
 
     public Film addLike(Long id, Long userId) {
+        if (!userStorage.checkUserExists(userId)) {
+            throw new NotFoundException("Пользователь с id = " + userId + " не найден");
+        }
         return storage.addLike(id, userId);
     }
 
     public Film deleteLike(Long id, Long userId) {
+        if (!userStorage.checkUserExists(userId)) {
+            throw new NotFoundException("Пользователь с id = " + userId + " не найден");
+        }
         return storage.deleteLike(id, userId);
     }
